requirements.txt
astroid==2.4.2
autopep8==1.5.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
pkg-resources==0.0.0
pycodestyle==2.6.0
pylint==2.6.0
pypng==0.0.20
six==1.15.0
toml==0.10.1
wrapt==1.12.1
